Code style: Inline `*_internal.h` headers in their .c files.
